bit o refactor
