+++ src/ffs/private/tree.ts
@@ -1,7 +1,8 @@
-import util from './util'
 import link from '../link'
+import util from '../util'
 import { PrivateTreeData, Tree, Links, File, PrivateTreeStatic, PrivateFileStatic, FileSystemVersion } from '../types'
 import { CID } from '../../ipfs'
+import keystore from '../../keystore'
 import PublicTree from '../public/tree'
 import PrivateFile from './file'
 import resolver from './resolver'
@@ -28,7 +29,7 @@ export class PrivateTree extends PublicTree {
   }
  
   static async empty(version: FileSystemVersion = FileSystemVersion.v1_0_0, key?: string): Promise<PrivateTree> {
-    const keyStr = key ? key : await util.genKeyStr()
+    const keyStr = key ? key : await keystore.genKeyStr()
     return new PrivateTree({}, version, keyStr)
   }
